+++ EagleSocial/Controller/MessageViewController.swift
@@ -9,32 +9,19 @@
 import UIKit
 import Firebase
 
-protocol CanRecieve {
-    func dataReceived(data: String)
-}
 
 class MessageViewController: UIViewController , UITableViewDelegate, UITableViewDataSource, UITextFieldDelegate {
 
     // Declare instance variables here
     var messageArray : [Message] = [Message]()
-    var delagate : CanRecieve?
     
     @IBOutlet var heightConstraint: NSLayoutConstraint!
     @IBOutlet var messageTextField: UITextField!
     @IBOutlet var conversationTableView: UITableView!
     @IBOutlet var sendButton: UIButton!
-    @IBOutlet var addNewUserToolBar: UIView!
-    @IBOutlet var addNewUserToolBarHeightConstraint: NSLayoutConstraint!
     
-    @IBOutlet var receipientLabel: UILabel!
-    @IBOutlet var receipientLabelBackgroundView: UIView!
-    @IBOutlet var addNewUserToConversationButton: UIButton!
-    
-    @IBOutlet var removeSelectedUserButton: UIButton!
-    
-    //TODO: - Relace with code to get conversationID from the Chat view controller.
-    //var conversationID : String = "-L8Y8gof4Ky_3ieWi6Ek"
-    var conversationID : String = ""
+    //TODO: - Relace with code to get conversationID from the Chat view controller. 
+    var conversationID : String = "-L8Y8gof4Ky_3ieWi6Ek"
     
     //Used to capture the keyboard height so that the messageText field
     // Will appear above the keyboard even on different devices of differen screen
@@ -69,19 +56,6 @@ class MessageViewController: UIViewController , UITableViewDelegate, UITableView
         //See function definition for more information on what this item does.
         configureTableView()
         
-        //Hide the Add New User Tool Bar if it is an existing
-        //conversation.
-        updateUserInterface()
-        
-        if conversationID != "" {
-            addNewUserToolBar.isHidden = true
-            addNewUserToolBarHeightConstraint.constant = 0
-        }
-        //Show the Add New User Tool bar if it is a new conversation.
-        else {
-            addNewUserToolBar.isHidden = false
-            
-        }
         //Retrieve messages upon loading the message screen.
         retrieveMessage()
 
@@ -105,7 +79,6 @@ class MessageViewController: UIViewController , UITableViewDelegate, UITableView
         messageCell.nameLabel.text = messageArray[indexPath.row].getSenderId()
         
         //Load the user's profile image into the profilImageView in the TableView Message Cell
-        //TODO: - Modify the classes/models to pull down user's profile picture.
         messageCell.profileImageView.image = UIImage(named: "profile_icon")
         
         //Post the newly created messageCell into the TableView.
@@ -143,7 +116,7 @@ class MessageViewController: UIViewController , UITableViewDelegate, UITableView
             
             //Construct the members list for the current conversation/message.
             let members : [String : Bool] = [String((Auth.auth().currentUser?.uid)!) : true,
-                                             String("IJHzyU11xmgk29y68TIhT7YkZHQ2") : true]
+                                             String("WzRw6ypWLvOAOo6iCd680Oz4Fwx1") : true]
             
             //Construct the message dictionary for the current conversation/message.
             //Sender, MessageBody, Conversation
@@ -155,7 +128,7 @@ class MessageViewController: UIViewController , UITableViewDelegate, UITableView
             let message = Message(membersa: members, messageDictionarya: messageDictionary)
             
             //Send the message.
-            self.conversationID = message.sendMessage()
+            message.sendMessage()
             
             //ReEnable the messageTextField so the user can compose more messages.
             self.messageTextField.isEnabled = true
@@ -174,33 +147,30 @@ class MessageViewController: UIViewController , UITableViewDelegate, UITableView
         //Retrive messages as they come in.
         //Make a call to the retrieveMessages method of the
         //Message class.
-        if conversationID != "" {
+        
+            let messageDB = Database.database().reference().child("Conversation").child(conversationID).child("Messages")
+            //let message1 = Message()
+        
+            messageDB.observe(.childAdded) { (snapshot) in
+                let snapshotValue = snapshot.value as! Dictionary<String,Any>
+                let message1 = Message()
             
-                let messageDB = Database.database().reference().child("Conversation").child(conversationID).child("Messages")
-                //let messageDB = Database.database().reference().child("Conversation").child("Messages")
-                //let message1 = Message()
+                message1.setMessageBody(messageBod: String(describing: snapshotValue["MessageBody"]!))
+                message1.setSenderId(sender: String(describing: snapshotValue["Sender"]!))
+                //message1.conversationID = self.conversationID
+                message1.setMessageDictionary(messageDict:  ["Sender": snapshotValue["Sender"]!,
+                                                         "MessageBody": snapshotValue["MessageBody"]!,
+                                                         "ConversationID" : self.conversationID])
+        
+                //return message1
             
-                messageDB.observe(.childAdded) { (snapshot) in
-                    let snapshotValue = snapshot.value as! Dictionary<String,Any>
-                    let message1 = Message()
-                
-                    message1.setMessageBody(messageBod: String(describing: snapshotValue["MessageBody"]!))
-                    message1.setSenderId(sender: String(describing: snapshotValue["Sender"]!))
-                    //message1.conversationID = self.conversationID
-                    message1.setMessageDictionary(messageDict:  ["Sender": snapshotValue["Sender"]!,
-                                                             "MessageBody": snapshotValue["MessageBody"]!,
-                                                             "ConversationID" : self.conversationID])
-            
-                    //return message1
-                
-                    //message.retrieveMessages(conversation: conversationID)
-            
-                    self.messageArray.append(message1)
-            
-                    self.configureTableView()
-            
-                    self.conversationTableView.reloadData()
-            }
+                //message.retrieveMessages(conversation: conversationID)
+        
+                self.messageArray.append(message1)
+        
+                self.configureTableView()
+        
+                self.conversationTableView.reloadData()
         }
     }
     /*
@@ -208,32 +178,12 @@ class MessageViewController: UIViewController , UITableViewDelegate, UITableView
     */
     @IBAction func backbtnPressed(_ sender: UIBarButtonItem) {
         
-        //clear the conversation ID property.
-        delagate?.dataReceived(data: "")
-        
         //Dissmiss the messages screen.
         //Return to the message list.
         dismiss(animated: true, completion: nil)
         
-       
-        
-    }
-    @IBAction func addNewUserToConversation(_ sender: Any) {
-        
-        performSegue(withIdentifier: "selectUserToChatWith", sender: self)
-        
-        receipientLabel.text = "Michael Pearson"
-        updateUserInterface()
-    }
-   
-    @IBAction func removeUserFromConversation(_ sender: Any) {
-        receipientLabel.text = ""
-        
-        //TODO: - clear the userID variable too.
-        updateUserInterface()
     }
     
-
     //Configure the conversationTableView properties here:
     func configureTableView() {
         //TODO: - Figureout why this line of code isn't working.
@@ -287,17 +237,4 @@ class MessageViewController: UIViewController , UITableViewDelegate, UITableView
         //Set the height constraint back to 50.
         self.heightConstraint.constant = CGFloat(50.0)
     }
-    func updateUserInterface() {
-     receipientLabelBackgroundView.layer.cornerRadius = 20
-        if receipientLabel.text != "" {
-            receipientLabelBackgroundView.backgroundColor = UIColor(hexString: "FFC14C")
-            addNewUserToConversationButton.isHidden = true
-            removeSelectedUserButton.isHidden = false
-        } else {
-            receipientLabelBackgroundView.backgroundColor = UIColor(named: "Clear")
-            addNewUserToConversationButton.isHidden = false
-            removeSelectedUserButton.isHidden = true
-        }
-    }
 }
-
