--- EagleSocial/ConversationTableViewCell.swift
@@ -1,24 +1,0 @@
-//
-//  ConversationTableViewCell.swift
-//  EagleSocial
-//
-//  Created by Michael Pearson on 3/3/18.
-//  Copyright © 2018 Jody Bailey. All rights reserved.
-//
-
-import UIKit
-
-class ConversationTableViewCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-    
-}
