+++ EagleSocial/Model/Message.swift
@@ -36,7 +36,7 @@ class Message {
     }
    
     //MARK: - Send Messages
-    func sendMessage() -> String
+    func sendMessage()
     {
         let conversationDB = Database.database().reference().child("Conversation")
         //let messagesDB = conversationDB.child("Messages")
@@ -44,13 +44,10 @@ class Message {
         //conversationID = conversationIDb
         
         //If this is a new conversation
-        if conversationID == "Optional(\"\")" || conversationID == "" {
+        if conversationID == "Optional(\"\")" {
             conversationID = conversationDB.childByAutoId().key
             messageDictionary["ConversationID"] = conversationID
-            let convo = Conversation(convoId: conversationID, mem: members, mes: messageDictionary)
-            convo.postNewConversation()
-            
-            /*conversationDB.child(conversationID).updateChildValues(["Members" : members]) {
+            conversationDB.child(conversationID).updateChildValues(["Members" : members]) {
                 (error, reference) in
                 
                 
@@ -58,25 +55,11 @@ class Message {
                     print(error!)
                     
                 } else {
-                    
-                    
                     conversationDB.child(self.conversationID).child("Messages").childByAutoId().updateChildValues(self.messageDictionary)
-                    
-                    print (self.conversationID + " Message saved successfully!!")
-                }
-            }
-            conversationDB.child(conversationID + "/LastMessage").setValue(messageDictionary["MessageBody"])
-            {
-                (error, reference) in
-                
-                if error != nil {
-                    print(error!)
-                    
-                } else {
                     
                     print ("Message saved successfully!!")
                 }
-            }*/
+            }
         }
         //If this is an existing conversation
         else  {
@@ -91,20 +74,8 @@ class Message {
                     print ("Message saved successfully!!")
                 }
             }
-            conversationDB.child(conversationID + "/LastMessage").setValue(messageDictionary["MessageBody"])
-            {
-                (error, reference) in
-                
-                if error != nil {
-                    print(error!)
-                    
-                } else {
-                    
-                    print ("Message saved successfully!!")
-                }
-            }
         }
-        return conversationID
+        
     }
     
     //MARK: - Receive Messages
