--- EagleSocial/Supporting Files/EagleSocial-Bridging-Header.h
@@ -1,8 +1,0 @@
-//
-//  Use this file to import your target's public headers that you would like to expose to Swift.
-//
-
-#import "SVProgressHUD.h"
-#import "SVRadialGradientLayer.h"
-#import "SVProgressAnimatedView.h"
-#import "SVIndefiniteAnimatedView.h"
