--- EagleSocial/Supporting Files/SVProgressHUD-Prefix.pch
@@ -1,7 +1,0 @@
-//
-// Prefix header for all source files of the 'SVProgressHUD' target in the 'SVProgressHUD' project
-//
-
-#ifdef __OBJC__
-    #import <Foundation/Foundation.h>
-#endif
