--- EagleSocial/View/SelectFriendTableViewCell/SelectFriendTableViewCell.xib
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="13771" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13772"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="SelectFriendTableViewCell" rowHeight="92" id="KGk-i7-Jjw" customClass="SelectFriendTableViewCell" customModule="EagleSocial" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="296" height="92"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="296" height="91.5"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="k0y-Xj-APS">
-                        <rect key="frame" x="70" y="5" width="221" height="81.5"/>
-                        <subviews>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Name" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="W8P-Ke-A8I">
-                                <rect key="frame" x="5" y="5" width="211" height="21"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                <nil key="textColor"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Major" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="kaj-yk-xjQ">
-                                <rect key="frame" x="5" y="39" width="211" height="21"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                <nil key="textColor"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstAttribute="trailing" secondItem="W8P-Ke-A8I" secondAttribute="trailing" constant="5" id="ChT-fv-flR"/>
-                            <constraint firstItem="W8P-Ke-A8I" firstAttribute="top" secondItem="k0y-Xj-APS" secondAttribute="top" constant="5" id="PlS-gC-s4a"/>
-                            <constraint firstItem="kaj-yk-xjQ" firstAttribute="leading" secondItem="k0y-Xj-APS" secondAttribute="leading" constant="5" id="Rgi-mK-7Vg"/>
-                            <constraint firstAttribute="trailing" secondItem="kaj-yk-xjQ" secondAttribute="trailing" constant="5" id="fmt-CW-uyr"/>
-                            <constraint firstItem="kaj-yk-xjQ" firstAttribute="top" secondItem="W8P-Ke-A8I" secondAttribute="bottom" constant="13" id="l8c-8l-rJB"/>
-                            <constraint firstItem="W8P-Ke-A8I" firstAttribute="leading" secondItem="k0y-Xj-APS" secondAttribute="leading" constant="5" id="tmq-ab-iVK"/>
-                        </constraints>
-                    </view>
-                    <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="2ni-35-cTt">
-                        <rect key="frame" x="5" y="16.5" width="60" height="60"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="60" id="aN4-RT-AdE"/>
-                            <constraint firstAttribute="width" constant="60" id="iiM-SR-YvT"/>
-                        </constraints>
-                    </imageView>
-                </subviews>
-                <constraints>
-                    <constraint firstAttribute="bottom" secondItem="k0y-Xj-APS" secondAttribute="bottom" constant="5" id="2ZL-a2-bmH"/>
-                    <constraint firstItem="2ni-35-cTt" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="5" id="Dpa-0F-rRt"/>
-                    <constraint firstItem="k0y-Xj-APS" firstAttribute="leading" secondItem="2ni-35-cTt" secondAttribute="trailing" constant="5" id="Rea-IA-ZMW"/>
-                    <constraint firstAttribute="trailing" secondItem="k0y-Xj-APS" secondAttribute="trailing" constant="5" id="XSX-Ji-JP3"/>
-                    <constraint firstAttribute="bottom" secondItem="2ni-35-cTt" secondAttribute="bottom" constant="15" id="YqH-EQ-3pw"/>
-                    <constraint firstItem="k0y-Xj-APS" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="5" id="jRc-Xp-ov4"/>
-                    <constraint firstItem="2ni-35-cTt" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="16.5" id="yKX-kF-imr"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <connections>
-                <outlet property="majorLabel" destination="kaj-yk-xjQ" id="2L3-Ag-LGh"/>
-                <outlet property="nameLabel" destination="W8P-Ke-A8I" id="0GY-0d-G0T"/>
-                <outlet property="profileImage" destination="2ni-35-cTt" id="FuQ-tr-K4X"/>
-            </connections>
-            <point key="canvasLocation" x="63" y="21"/>
-        </tableViewCell>
-    </objects>
-</document>
